refactor(exchange): share simplifyError across brokers
Moves simplifyError to src/util/ and attaches it to the Trading212 client
so its rejected errors also drop the verbose Axios internals that can leak
API credentials.
