+++ packages/exchange/src/broker/alpaca/api/AlpacaAPI.ts
@@ -6,7 +6,7 @@ import {BarsResponseSchema, LatestBarsResponseSchema} from './schema/BarSchema.j
 import {ClockSchema} from './schema/ClockSchema.js';
 import {OrderSchema, type AlpacaAssetClass} from './schema/OrderSchema.js';
 import {PositionSchema} from './schema/PositionSchema.js';
-import {simplifyError} from './simplifyError.js';
+import {simplifyError} from '../../../util/simplifyError.js';
 import {SnapshotsResponseSchema} from './schema/SnapshotSchema.js';
 import {z} from 'zod';
