+++ packages/exchange/src/util/simplifyError.test.ts
